refactored to rename 'helpers' to 'hooks'
