+++ apps/ds-ui/src/components/prayer/PrayerSettings.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 import { Form } from 'react-bootstrap';
-import { useUserSettings } from '../../helpers/UserSettings';
+import { useUserSettings } from '../../hooks/UserSettings';
 import { LoadingMessage, ErrorLoadingDataMessage } from '../common/loading';
 import { UserAttributes } from '@devouringscripture/common';
 import { PrayerViewFilterComponent } from './PrayerViewFilterComponent';
